backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
langchain-openai
langchain-community
qdrant-client
sentence-transformers
fuzzywuzzy
python-Levenshtein
httpx